added project common + project page switch
