+++ src/main/java/org/app/controllers/HomePageController.java
@@ -1,15 +1,9 @@
 package org.app.controllers;
 
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.app.controllers.util.ProjectCommon;
 import org.app.exceptions.UserNotFoundException;
-import org.app.model.Board;
-import org.app.model.Project;
-import org.app.model.User;
-import org.app.service.ProjectService;
-import org.app.service.TodoTaskService;
-import org.app.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -21,32 +15,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/")
 public class HomePageController {
 
-    private final UserService userService;
-    private final TodoTaskService taskService;
-    private final ProjectService projectService;
+    private final ProjectCommon projectCommon;
 
     @GetMapping
     public String showHomePage(Model model)
         throws UserNotFoundException {
-        User currentUser = userService.getCurrentUser();
-        if (currentUser == null) {
-            return "redirect:/login";
-        } else if (currentUser.getProjects().isEmpty()) {
-            return "redirect:/project/show";
-        } else {
-            List<Project> projects = projectService.getAllProjects();
-            Project activeProject = projects.get(0);
-            List<Board> boards = activeProject.getBoards();
-            int totalTasks = taskService.getTasksByUserId(currentUser.getId()).size();
-            int completedTasks = taskService.getCompletedTasksCount(currentUser.getId());
-            model.addAttribute("activeProject", activeProject);
-            model.addAttribute("userBoards", boards);
-            model.addAttribute("currentUser", currentUser);
-            model.addAttribute("totalTasks", totalTasks);
-            model.addAttribute("completedTasks", completedTasks);
-            model.addAttribute("allProjects", projects);
-            return "home";
-        }
+       return projectCommon.getHomePageUtility(model, null);
     }
 
 }
