+++ src/main/java/org/app/model/User.java
@@ -92,4 +92,14 @@ public class User implements UserDetails {
     public boolean isEnabled() {
         return true;
     }
+
+    @Override
+    public String toString() {
+        return "User{" +
+            "id=" + id +
+            ", email='" + email + '\'' +
+            ", name='" + name + '\'' +
+            ", username='" + username + '\'' +
+            '}';
+    }
 }
