+++ src/main/resources/templates/fragments/navbar-fragment.html
@@ -37,11 +37,12 @@
       </form>
     </div>
 
-    <div id="projects-slider" class="slide-left"  th:each="project : ${allProjects}" th:attr="data-board-id=${project.projectId}">
+    <div id="projects-slider" class="slide-left">
       <a href="javascript:void(0)" class="close-btn" id="projSlider-close">&times;</a>
-      <a href="#" th:text="${project.projectName}"></a>
-
+      <a th:href="@{/project/find/{projectId}(projectId = ${project.projectId})}"
+         th:text="${project.projectName}" th:each="project : ${allProjects}"></a>
     </div>
+
   </div>
