+++ src/main/resources/templates/fragments/sidebar-fragment.html
@@ -8,7 +8,7 @@
     <div class="sidebar">
         <div class="sidebar-menu">
             <div>
-                <h1 th:value="${project}"></h1>
+                <h3 th:text="${activeProject.projectName}"></h3>
             </div>
             <a sec:authorize="isAuthenticated()">Project</a>
             <ul sec:authorize="isAuthenticated()">
